Add low voltage and over voltage protection
This part is a final safety net for battery charage management.

It is configured to have a low voltage cutoff of 2.5V and an over
voltage cutoff of 4.3V.

These values are comfortably outside the range of the change controller
thresholds, but still within the ranges generally accepted to avoid
permanent cell damage.

If for some reason we over charge or over discharge, this part will do a
low side disconnect of the battery from the rest of the system.

lib: Add simple logos

Currently unused, but are available.

See:
  * https://meshtastic.org/docs/legal/licensing-and-trademark/#community-project-logos-m-powered--m-pwrd
  * https://oshwa.org/resources/open-source-hardware-logo/